split the berryclient into restsolver and streamingsolver
