+++ cloudberry/zion/src/main/scala/edu/uci/ics/cloudberry/zion/model/datastore/IPostTransform.scala
@@ -1,0 +1,13 @@
+package edu.uci.ics.cloudberry.zion.model.datastore
+
+import play.api.libs.json.JsValue
+
+
+trait IPostTransform {
+  def transform(jsValue: JsValue): JsValue
+}
+
+case object NoTransform extends IPostTransform {
+  override def transform(jsValue: JsValue): JsValue = jsValue
+}
+
